Refactor due to PR suggestions
